Eliminate cast

+++ main/php_scandir.c
@@ -73,7 +73,7 @@ PHPAPI int php_scandir(const char *dirname, struct dirent **namelist[], int (*se
 	}
 
 	while (!php_readdir_r(dirp, (struct dirent *)entry, &dp) && dp) {
-		int dsize = 0;
+		size_t dsize = 0;
 		struct dirent *newdp = NULL;
 
 		if (selector && (*selector)(dp) == 0) {
@@ -95,7 +95,7 @@ PHPAPI int php_scandir(const char *dirname, struct dirent **namelist[], int (*se
 			vector = newv;
 		}
 
-		dsize = sizeof (struct dirent) + (((int)strlen(dp->d_name) + 1) * sizeof(char));
+		dsize = sizeof (struct dirent) + ((strlen(dp->d_name) + 1) * sizeof(char));
 		newdp = (struct dirent *) malloc(dsize);
 
 		if (newdp == NULL) {
